+++ TP1/conversionTest.c
@@ -3,24 +3,24 @@
 #include <assert.h>
 
 int main() {
-    float lower = -300;
-    float upper =  300;
-    float step  =   20;
+    double lower = -300;
+    double upper =  300;
+    double step  =   20;
     //Defino yo lower y upper, aunque tambien podria hacer que el usuario ingrese los valores que quiera. 
 
     assert(lower <= upper); //Si el upper es menor al lower me tira error
 
     printf("Testing Fahrenheit to Celsius conversion:\n");
-    for (float fahr = lower; fahr <= upper; fahr += step) {
-        float celsius = fahrenheitToCelsius(fahr);
+    for (double fahr = lower; fahr <= upper; fahr += step) {
+        double celsius = fahrenheitToCelsius(fahr);
         printf("%6.1f\t= %6.1f\n", fahr, celsius);
     }
 
     printf("\n");
 
     printf("Testing Celsius to Fahrenheit conversion:\n");
-    for (float celsius = lower; celsius <= upper; celsius += step) {
-        float fahr = celsiusToFahrenheit(celsius);
+    for (double celsius = lower; celsius <= upper; celsius += step) {
+        double fahr = celsiusToFahrenheit(celsius);
         printf("%6.1f\t= %6.1f\n", celsius, fahr);
     }
